Logic for save review; template to display errors on failed save
Note that there are problems with the approach of rendering a route from
another controller and trying to pass in the save errors to render. The URL
is not stirctly correct, and the failed save content will render (though it disappears
when navigating away and returning). Propose storing errors in the current_user
object and/or being creative with Ruby callbacks to serve errors in the correct
part of the create Review life cycle.